+++ docs-portal/docusaurus.config.js
@@ -10,6 +10,10 @@ const config = {
     defaultLocale: "en",
     locales: ["en"]
   },
+  themes: ["@docusaurus/theme-mermaid"],
+  markdown: {
+    mermaid: true
+  },
   presets: [
     [
       "classic",